+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.4.2
+* fix issues(#76) support android 13.
+## 0.4.1
+* fix issues(#71)
+## 0.4.0
+* add silent install, fix google play release issues.
+## 0.3.9
+* fixed download file not exist error.
 ## 0.3.8+2
 * fixed download file not exist error.
 ## 0.3.8+1